compiling on windows

+++ include/et/rendering/base/vertexstorage.h
@@ -32,8 +32,8 @@ namespace et
 		{
 			ET_ASSERT(hasAttributeWithType(usage, T));
 			
-			return VertexDataAccessor<T>(data().binary(), data().dataSize(), stride(),
-				offset * stride() + offsetOfAttribute(usage));
+			return VertexDataAccessor<T>(data().binary(), static_cast<uint32_t>(data().dataSize()), 
+				stride(), offset * stride() + offsetOfAttribute(usage));
 		}
 		
 		template <DataType T>
